Fixed error in parsing floats starting with "-0."
Floats were made up of parsed integers with more components, and so
something like "-0.5e2" would be parsed as "-0" with an additional
".5e2", and then put back together after being matched and parsed as a
float. The problem was that the integer part would be converted to an
integer before things were put back. In the above example, the "-0"
would become a zero, and the end result would be "0.5e2".